add more sfx (singularity crafting, leave, enter) + move all dream listener
