+++ src/main/java/fr/openmc/core/features/city/view/CityViewManager.java
@@ -186,8 +186,8 @@ public class CityViewManager {
         particleLocations.forEach(location ->
                 ParticleUtils.sendParticlePacket(
                         player,
-                        location,
                         particle,
+                        location,
                         1,
                         0D, 0D, 0D,
                         0D,
