+++ src/main/java/fr/openmc/core/features/dream/DreamManager.java
@@ -18,12 +18,12 @@ import fr.openmc.core.features.dream.generation.DreamDimensionManager;
 import fr.openmc.core.features.dream.generation.listeners.CloudStructureDispenserListener;
 import fr.openmc.core.features.dream.generation.listeners.ReplaceBlockListener;
 import fr.openmc.core.features.dream.generation.structures.DreamStructuresManager;
-import fr.openmc.core.features.dream.listeners.biomes.PlayerEnteredBiome;
+import fr.openmc.core.features.dream.listeners.dream.PlayerEnteredBiome;
 import fr.openmc.core.features.dream.listeners.dream.*;
-import fr.openmc.core.features.dream.listeners.orb.PlayerObtainOrb;
-import fr.openmc.core.features.dream.listeners.others.CraftingConvertorListener;
-import fr.openmc.core.features.dream.listeners.others.PlayerEatSomnifere;
-import fr.openmc.core.features.dream.listeners.others.SingularityCraftListener;
+import fr.openmc.core.features.dream.listeners.dream.PlayerObtainOrb;
+import fr.openmc.core.features.dream.listeners.registry.CraftingConvertorListener;
+import fr.openmc.core.features.dream.listeners.dream.PlayerEatSomnifere;
+import fr.openmc.core.features.dream.mecanism.singularity.SingularityCraftListener;
 import fr.openmc.core.features.dream.listeners.registry.DreamItemEquipListener;
 import fr.openmc.core.features.dream.mecanism.cloudfishing.CloudFishingManager;
 import fr.openmc.core.features.dream.mecanism.cold.ColdManager;
@@ -43,7 +43,6 @@ import org.bukkit.Location;
 import org.bukkit.World;
 import org.bukkit.block.Biome;
 import org.bukkit.entity.Player;
-import org.bukkit.event.inventory.InventoryCloseEvent;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.PlayerInventory;
