+++ src/main/java/fr/openmc/core/features/dream/registries/mobs/Breezy.java
@@ -84,8 +84,8 @@ public class Breezy extends DreamMob implements Listener {
 
         ParticleUtils.sendParticlePacket(
                 target,
-                eye,
                 Particle.CLOUD,
+                eye,
                 20, 0.2, 0.2, 0.2, 0.01, null
         );
     }
